+++ packages/@webex/plugin-meetings/test/unit/spec/locus-info/index.js
@@ -10,6 +10,7 @@ import SelfUtils from '@webex/plugin-meetings/src/locus-info/selfUtils';
 import InfoUtils from '@webex/plugin-meetings/src/locus-info/infoUtils';
 import EmbeddedAppsUtils from '@webex/plugin-meetings/src/locus-info/embeddedAppsUtils';
 import LocusDeltaParser from '@webex/plugin-meetings/src/locus-info/parser';
+import Metrics from '@webex/plugin-meetings/src/metrics';
 
 import {
   LOCUSINFO,
@@ -24,7 +25,6 @@ import {
 } from '../../../../src/constants';
 
 import {self, selfWithInactivity} from './selfConstant';
-import uuid from 'uuid';
 
 describe('plugin-meetings', () => {
   describe('LocusInfo index', () => {
@@ -39,6 +39,7 @@ describe('plugin-meetings', () => {
     const locus = {};
     const meetingId = 'meetingId';
     let locusInfo;
+    let sendBehavioralMetricStub;
 
     const webex = new MockWebex({
       children: {
@@ -65,6 +66,12 @@ describe('plugin-meetings', () => {
           },
         },
       };
+
+      sendBehavioralMetricStub = sinon.stub(Metrics, 'sendBehavioralMetric');
+    });
+
+    afterEach(() => {
+      sinon.restore();
     });
 
     describe('#updateControls', () => {
@@ -1877,6 +1884,114 @@ describe('plugin-meetings', () => {
         });
       });
 
+      describe('edge cases for sync failing', () => {
+        const {DESYNC} = LocusDeltaParser.loci;
+        const fakeFullLocusDto = {id: 'fake full locus dto'};
+        let meeting;
+
+        beforeEach(() => {
+          sinon.stub(locusInfo.locusParser, 'resume');
+          sinon.stub(webex.meetings, 'destroy');
+
+          meeting = {
+            meetingRequest: {
+              getLocusDTO: sandbox.stub(),
+            },
+            locusInfo: {
+              handleLocusDelta: sandbox.stub(),
+              onFullLocus: sandbox.stub(),
+            },
+            locusUrl: 'fullSyncUrl',
+          };
+
+          locusInfo.locusParser.workingCopy = {
+            syncUrl: 'deltaSyncUrl',
+          };
+        });
+
+        it('applyLocusDeltaData gets full locus on DESYNC action if we do not have a syncUrl and destroys the meeting if that fails', () => {
+          meeting.meetingRequest.getLocusDTO.rejects(new Error('fake error'));
+
+          locusInfo.locusParser.workingCopy = {}; // no syncUrl
+
+          // Since we have a promise inside a function we want to test that's not returned,
+          // we will wait and stub it's last function to resolve this waiting promise.
+          return new Promise((resolve) => {
+            webex.meetings.destroy.callsFake(() => resolve());
+            locusInfo.applyLocusDeltaData(DESYNC, fakeLocus, meeting);
+          }).then(() => {
+            assert.calledOnceWithExactly(meeting.meetingRequest.getLocusDTO, {url: 'fullSyncUrl'});
+
+            assert.notCalled(meeting.locusInfo.handleLocusDelta);
+            assert.notCalled(meeting.locusInfo.onFullLocus);
+            assert.notCalled(locusInfo.locusParser.resume);
+
+            assert.calledOnceWithExactly(webex.meetings.destroy, meeting, 'LOCUS_DTO_SYNC_FAILED');
+          });
+        });
+
+        it('applyLocusDeltaData first tries a delta sync on DESYNC action and if that fails, does a full locus sync', () => {
+          meeting.meetingRequest.getLocusDTO.onCall(0).rejects(new Error('fake error'));
+          meeting.meetingRequest.getLocusDTO.onCall(1).resolves({body: fakeFullLocusDto});
+
+          // Since we have a promise inside a function we want to test that's not returned,
+          // we will wait and stub it's last function to resolve this waiting promise.
+          return new Promise((resolve) => {
+            locusInfo.locusParser.resume.callsFake(() => resolve());
+            locusInfo.applyLocusDeltaData(DESYNC, fakeLocus, meeting);
+          }).then(() => {
+            assert.calledTwice(meeting.meetingRequest.getLocusDTO);
+
+            assert.deepEqual(meeting.meetingRequest.getLocusDTO.getCalls()[0].args, [{url: 'deltaSyncUrl'}]);
+            assert.deepEqual(meeting.meetingRequest.getLocusDTO.getCalls()[1].args, [{url: 'fullSyncUrl'}]);
+
+            assert.calledWith(sendBehavioralMetricStub, 'js_sdk_locus_delta_sync_failed', {
+              correlationId: meeting.correlationId,
+              url: 'deltaSyncUrl',
+              reason: 'fake error',
+              errorName: 'Error',
+              stack: sinon.match.any,
+              code: sinon.match.any,
+            });
+
+            assert.notCalled(meeting.locusInfo.handleLocusDelta);
+            assert.calledOnceWithExactly(meeting.locusInfo.onFullLocus, fakeFullLocusDto);
+            assert.calledOnce(locusInfo.locusParser.resume);
+          });
+        });
+
+        it('applyLocusDeltaData destroys the meeting if both delta sync and full sync fail', () => {
+          meeting.meetingRequest.getLocusDTO.rejects(new Error('fake error'));
+
+          // Since we have a promise inside a function we want to test that's not returned,
+          // we will wait and stub it's last function to resolve this waiting promise.
+          return new Promise((resolve) => {
+            webex.meetings.destroy.callsFake(() => resolve());
+            locusInfo.applyLocusDeltaData(DESYNC, fakeLocus, meeting);
+          }).then(() => {
+            assert.calledTwice(meeting.meetingRequest.getLocusDTO);
+
+            assert.deepEqual(meeting.meetingRequest.getLocusDTO.getCalls()[0].args, [{url: 'deltaSyncUrl'}]);
+            assert.deepEqual(meeting.meetingRequest.getLocusDTO.getCalls()[1].args, [{url: 'fullSyncUrl'}]);
+
+            assert.calledWith(sendBehavioralMetricStub, 'js_sdk_locus_delta_sync_failed', {
+              correlationId: meeting.correlationId,
+              url: 'deltaSyncUrl',
+              reason: 'fake error',
+              errorName: 'Error',
+              stack: sinon.match.any,
+              code: sinon.match.any,
+            });
+
+            assert.notCalled(meeting.locusInfo.handleLocusDelta);
+            assert.notCalled(meeting.locusInfo.onFullLocus);
+            assert.notCalled(locusInfo.locusParser.resume);
+
+            assert.calledOnceWithExactly(webex.meetings.destroy, meeting, 'LOCUS_DTO_SYNC_FAILED');
+          });
+        });
+      });
+
       it('onDeltaLocus handle delta data', () => {
         fakeLocus.participants = {};
         const fakeBreakout = {
@@ -2419,6 +2534,8 @@ describe('plugin-meetings', () => {
 
         // send an out-of-order delta
         locusInfo.handleLocusDelta(oooDelta, mockMeeting);
+
+        assert.calledOnceWithExactly(sendBehavioralMetricStub, 'js_sdk_locus_delta_ooo', { stack: sinon.match.any})
 
         await clock.tickAsync(12499);
         await testUtils.flushPromises();